feat: use spawn on long ffmpeg commands
The spawn command lets us process data on-the-fly as it
becomes available. This lets us track progress, to prevent
the watchdog from killing the service, as well as preventing
max limit on cache for the exec command, which will cache
all output, and turn it to a big string, which can crash.

Also a lot of refactoring done.